+++ Vault-Coder/Changes.md
@@ -8,6 +8,8 @@
 - Merging now supports annotations (e.g. `@deprecated`), whereas earlier versions omitted them from merge results
 - Functional parameters that accepted tuples were not parsed correctly when merging - fixed
 ### New Features
+- Added a new application / mode that writes input model templates based on existing database structure 
+  (see README for details)
 - Added `-N` argument that disables merging
 - Added access point -writing for combined models
 - Added FromModelFactory writing for the data models (experimental)